+++ app/src/main/assets/CHANGELOG.md
@@ -6,9 +6,10 @@ For support you can use [the contact form](https://contact.faircode.eu/?product=
 
 ### [Ubirajara](https://en.wikipedia.org/wiki/Ubirajara_jubatus)
 
-### Next version
+### 1.2094 - 2023-08-14
 
 * Added rule action to add local notes
+* Removed a workaround for an Android issue that caused side effects
 * Updated [translations](https://crowdin.com/project/open-source-email)
 
 ### 1.2093 - 2023-08-14